+++ www/go/core/views/extjs3/form/EntityPanel.js
@@ -88,6 +88,10 @@ go.form.EntityPanel = Ext.extend(Ext.form.FormPanel, {
 		Ext.apply(v, this.getForm().getFieldValues(dirtyOnly));
 		return v;
 	},
+
+	isDirty : function() {
+		return this.getForm().isDirty();
+	},
 	
 	setValues : function(v, trackReset) {
 		var field, name;
@@ -171,6 +175,8 @@ go.form.EntityPanel = Ext.extend(Ext.form.FormPanel, {
 
 				me.fireEvent("submit", me, true, serverId);
 
+				me.form.trackReset();
+
 				return me.entityStore.single(serverId).then(function(entity) {
 					me.entity = entity;
 					me.currentId = serverId;
@@ -229,6 +235,8 @@ go.form.EntityPanel = Ext.extend(Ext.form.FormPanel, {
 		})
 
 	},
+
+
 
 	markServerValidationErrors : function(e, fieldPrefix) {
 		var firstError;