+++ www/go/core/views/extjs3/form/FormGroup.js
@@ -283,6 +283,8 @@ go.form.FormGroup = Ext.extend(Ext.Panel, {
 		this.focusNewField(wrap);
 		//wrap.formField.focus();
 
+		this.dirty = true;
+
 		this.fireEvent("newitem", this, wrap);
 	},
 	
@@ -461,10 +463,30 @@ go.form.FormGroup = Ext.extend(Ext.Panel, {
 		return this.name;
 	},
 
+
+	/**
+	 * Required for  resetting after loading a form
+	 */
+	setNotDirty : function() {
+
+		var fn = function (i) {
+			i.originalValue = i.getValue();
+			i.dirty = false;
+			if(i.setNotDirty) {
+				i.setNotDirty(false);
+			}
+		};
+		this.getAllFormFields().forEach(fn, this);
+	},
+
 	
 	isDirty: function () {
 		if(this.dirty) {
 			return true;
+		}
+
+		if(!this.items) {
+			return false;
 		}
 
 		var dirty = false;
@@ -488,7 +510,7 @@ go.form.FormGroup = Ext.extend(Ext.Panel, {
 	},
 
 	setValue: function (records) {
-		this.dirty = true;
+		// this.dirty = true;
 		this.removeAll();
 		if(records === null) return;
 		this.markDeleted = [];